% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate_pairs.R
\name{get_unique_pairs}
\alias{get_unique_pairs}
\title{Get unique pairs from bulk fastq files}
\usage{
get_unique_pairs(fastq_id1s)
}
\arguments{
\item{fastq_id1s}{Character vector of first sequence identifiers from
fastq.gz files. Returned from \code{\link{get_fastq_id1s}}.}
}
\value{
Character vector of unique pairs.
}
\description{
Get unique pairs from bulk fastq files
}
