% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_seq.R
\name{get_vsd}
\alias{get_vsd}
\title{Get variance stabilized data for exploratory data analysis}
\usage{
get_vsd(eset)
}
\arguments{
\item{eset}{ExpressionSet loaded with \link{load_seq}.}
}
\value{
\code{matrix} with variance stabilized expression data.
}
\description{
Get variance stabilized data for exploratory data analysis
}
